Reworking how the image upload and resizing work is done (SCHEMA).
Updated schema, as a previous regression has broken the initial intention of this feature.

Now a single image media object is saved, with each of the 3 images attached to it.  Previously it was one, and by convention the thumbnail and resized were located in the same folder and not mapped in the database.  A regression broke this and 3 media objects were getting added to the database and they were all in there own folder.  This regression was never the intention but to fix it properly, I've made some improvements on the schema so we no longer rely on convention that can be easily broken with a regression.

Unit tests should be included but have been omitted for now.  This is prolonged technical debt as this application needs some serious attention towards unit testing and integration testing.