Implemented media object upload (not using resumable) for admin API uploads.
The designed use case for this is asset conversion from project to project.
This is work in progress towards the final goal of this application fulfilling the requirements of our conversion from cloud media providers to us.
Another commit will allow after an upload, to convert old scene.scene.$.urls to be changed from a cloud provider to us.

We may need to improve the video object upload to optionally allow a list of files that includes the transcoded results of a original video media file.

I've deprecated the old tests and began with a fresh slate.  These should be greatly expanded when possible.
Resumable API pretty difficult to test without a browser doing the upload so we will need to work on this point.